My first R side project
This project was done to start to explore how to prepare a dataset for analysis and visualization.  This data was originally copied from page 26 of this report...
http://nebula.wsimg.com/1a7f83c14af6a516176740244d8afc46?AccessKeyId=DAC3A56D8FB782449D2A&disposition=0&alloworigin=1

I have added both the original copied dataset and the manually edited one for those who might like to take a crack at parsing it programmatically (this report does come out annually).